+++ source/Axiom/Axiom/Codecs/Subtitle/Copy.cs
@@ -1,0 +1,76 @@
+﻿/* ----------------------------------------------------------------------
+Axiom UI
+Copyright (C) 2017-2019 Matt McManis
+http://github.com/MattMcManis/Axiom
+http://axiomui.github.io
+mattmcmanis@outlook.com
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.If not, see <http://www.gnu.org/licenses/>. 
+---------------------------------------------------------------------- */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Axiom
+{
+    public class SubtitleCopy
+    {
+        // -------------------------
+        // Codec
+        // -------------------------
+        public static string codec = "-c:s copy";
+
+        // -------------------------
+        // Stream
+        // -------------------------
+        // selected item
+        public static string stream = "all";
+
+        // -------------------------
+        // Checked
+        // -------------------------
+        public static void controlsChecked(ViewModel vm)
+        {
+            // None
+        }
+
+        // -------------------------
+        // Enabled
+        // -------------------------
+        public static void controlsEnable(ViewModel vm)
+        {
+            // Subtitle Codec
+            vm.SubtitleCodec_IsEnabled = true;
+
+            // Subtitle Stream
+            vm.SubtitleStream_IsEnabled = true;
+
+            // Subtitle List View
+            // Controlled in cboSubtitleStream_SelectionChanged
+        }
+
+        // -------------------------
+        // Disabled
+        // -------------------------
+        public static void controlsDisable(ViewModel vm)
+        {
+            // None
+        }
+
+
+    }
+}